% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legends.R
\name{inset_themes}
\alias{inset_themes}
\alias{inset_legend_dark}
\alias{inset_legend_light}
\title{inset map legend themes}
\usage{
inset_legend_dark(location = "top-left")

inset_legend_light(location = "top-left")
}
\arguments{
\item{location}{the location to inset legends. One of "top-left", "top-right", "bottom-left", "bottom-right"}
}
\value{
\code{inlegend_dark} returns dark legend theme based on kepler.gl as an object of classes \code{theme} and \code{gg}.

\code{inlegend_light} returns a light white legend theme as an object of classes \code{theme} and \code{gg}.
}
\description{
Inset legend themes for ggplot2 maps
}
\details{
These themes move and restyle the ggplot2 legend to be more in line with mapping nroms.
\code{ggplot2} are inset in the desired map corner. The legends are restyled to appear simpler and draw less attention when overlaid on maps.
}
\author{
Miles McBain
}
