% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cropped_map.R
\name{theme_cropped_map}
\alias{theme_cropped_map}
\title{theme_cropped_map}
\usage{
theme_cropped_map(...)
}
\arguments{
\item{...}{arguments passed to \code{ggplot2::theme()}}
}
\value{

}
\description{
Remove plot axis and whitespace margins
}
\details{
In a map where the legend is inset it may be convenient to have the plot
tightly cropped around the map extent. This theme give you a such a plot.
Beware that there is no space for titles, so that kind of information must
be conveyed in section headings or captions.

When saving map using this theme as an image, you will see whitespace in the
output unless the aspect ratio of your output image matches the aspect ratio
of the map (plot area) exactly.
}
\author{
Miles McBain
}
